Move page search view to wagtail.admin.views.pages.search
